reconfigured files for proper lambda execution
