+++ data_processing.py
@@ -1,82 +1,46 @@
-import typing
-import json
+import boto3
+
 from modes import overview
 from modes import resolvers
 
 
-with open("data.json", "r") as read_file:
-  data = json.load(read_file)
+def lambda_handler(event: dict, context: dict) -> dict:
+    """
+    Main handler for the DataProcessing lambda.
 
-def get_overview_data(data: dict) -> dict:
-  """
-  Process data for overview component.
+      Makes connection to dynamoDB.
+      Passes API key as event to dynamo and returns raw data.
+      Processes raw data and returns to calling context.
 
-  Args:
-    data: input data as a dictionary.
+    Args:
+      event:         dict containing API key.
+      context:       dict containing calling context info.
+    
+    Returns:
+      processedData: dict containing processed data
+    """
+    # Make a connection to dynamodb service with boto3.
+    client = boto3.client('dynamodb')
 
-  Returns:
-    p_dic: processed data as a dictionary.
-  """
-  # Init processed data dictionary.
-  p_dic: dict = {
-    'summary': overview.get_summary(data),
-    'requests': overview.get_requests(data),
-    'response': overview.get_responses(data),
-    'resolvers': overview.get_resolvers(data)
-  }
+    # getItem api call, pass in params
+    response = client.get_item(
+        TableName = "UserData",
+        Key = {
+            'UserID': {
+                'S': event['UserID']
+            }
+        }
+    )
 
-  # Return processed data.
-  return p_dic
+    # Extract data from the response.
+    data = response
 
+    # Process data.
+    processedData: dict = {
+      'overview': overview.get_overview_data(data),
+      'resolvers': resolvers.get_resolvers_data(data)
+    }
+    
+    # Return
+    return processedData
 
-def get_resolvers_data(data: dict) -> dict:
-  """
-  Processes data for resolvers mode.
-
-  Args:
-    data: a dictionary of data for processing.
-
-  Returns:
-    p_dic: a dictionary of component-level data.
-  """
-
-  # Init processed data dictionary.
-  p_dic: dict = {
-    "invocationCounts": resolvers.get_invocationcounts(data),
-    "executionTimes": resolvers.get_aveexecutiontimes(data),
-  }
-
-  # Stick some averages data on p_dic.
-  total_count = sum([p_dic['invocationCounts'][x]
-                     for x in p_dic['invocationCounts']])
-  sum_ave = \
-    sum([p_dic['executionTimes'][x] * p_dic['invocationCounts'][x]
-         for x in p_dic['invocationCounts']])
-  p_dic['averageTime'] = sum_ave / total_count
-
-  # Return the processed data.
-  return p_dic
-
-
-# Process the data.
-
-def main(data: dict) -> None:
-
-  processedData: dict = {
-    'overview': get_overview_data(data),
-    'resolvers': get_resolvers_data(data)
-  }
-
-  file_object = open('processed_data.json', 'w')
-
-  # Save dict data into the JSON file.
-  json.dump(processedData, file_object)
-
-  return
-
-
-
-#print(processedData)
-
-if __name__ == "__main__":
-  main(data)
